==> oligarchies/1_data_clean.ipynb <==
import pandas as pd
import networkx as nx

df = pd.read_csv('../data_light/authors_journals.csv') \
       .drop(['Unnamed: 0','aff_name', 'aff_id','aff_city_id','author_name', 'aff_country_code'], axis=1) \
       .rename(columns={'journal.title': 'journal'}) 

# merge with pub_info to get the year
df = df.merge(pd.read_csv('../data_light/pubs_info.csv'), on='pub_id')

# merge with author_info to get the gender
df = df.merge(pd.read_csv('../data_light/authors_info.csv').drop(['current_organization_id'], axis=1), on='researcher_id')

# just a pub_id - research_id pair, no duplicated, groupby
df = df.groupby(['pub_id', 'researcher_id']).agg({'journal': 'first',
                                                  'year': 'first',
                                                  'gender': 'first',
                                                  'LMIC': 'max'}).reset_index()

# drop PLOS Medicine
df = df[df['journal'] != 'PLOS Medicine']

# drop year 2023
df = df[df['year'] != 2023]

# exclude papers with only one author
df1 = df.groupby('pub_id').filter(lambda x: len(x) > 1) 
print(f"{len(df.researcher_id.unique())} authors (beginning)")
print(f"{len(df.pub_id.unique())} publications (beginning)")
print(f"{len(df.researcher_id.unique()) - len(df1.researcher_id.unique())} authors dropped where there was only one author")
print(f"{len(df1.researcher_id.unique())} authors")


# exclude papers with more than 100 authors
df2 = df1.groupby('pub_id').filter(lambda x: len(x) <= 100)
print(f"{len(df1.researcher_id.unique()) - len(df2.researcher_id.unique())} authors dropped where papers had more than 100 authors")
print(f"{len(df2.researcher_id.unique())} authors")


# drop if gender is unknown
df3 = df2[df2.gender.isna() == False]
print(f"{len(df2.researcher_id.unique()) - len(df3.researcher_id.unique())} authors dropped with no gender info")
print(f"{len(df3.researcher_id.unique())} authors (final)")
print(f"{len(df3.pub_id.unique())} publications (final)")

# transform female into 1, male into 0
df3['gender'] = df3['gender'].apply(lambda x: 1 if x == "female" else 0)

df = df3.copy()

df.to_csv('../data_light/data_clean.csv', index=False)

df.head()

# count na 
df3.isna().sum()

# count the number of authors in total
print(f"Total number of authors: {len(df.researcher_id.unique())}")
print(f"Total number of publications: {len(df.pub_id.unique())}")

# count the number of authors per publication, sort desc
af = df.groupby('pub_id') \
       .agg({'researcher_id': 'count'}) \
       .reset_index() \
       .sort_values('researcher_id', ascending=False)

# print median and IQR of the number of authors per publication
print(f"Median number of authors per publication: {af.researcher_id.median()}")
print(f"IQR of the number of authors per publication: {af.researcher_id.quantile(0.25) , af.researcher_id.quantile(0.75)}")


df[df.year == 2007].pub_id.nunique()

# print the number of publications per year, in bins of 4 years
# 2007-2010, 2011-2014, 2015-2018, 2019-2022
# drop 2023
df = df[df.year != 2023]

a = df[df.year.isin([2007, 2008, 2009, 2010])].pub_id.nunique()
b = df[df.year.isin([2011, 2012, 2013, 2014])].pub_id.nunique()
c = df[df.year.isin([2015, 2016, 2017, 2018])].pub_id.nunique()
d = df[df.year.isin([2019, 2020, 2021, 2022])].pub_id.nunique()

total_pubs = df.pub_id.nunique()

# sum of publications in bins of 4 years
print(f"Number of publications in 2007-2010: {a} ({a*100/total_pubs:.1f})")
print(f"Number of publications in 2011-2014: {b} ({b*100/total_pubs:.1f})")
print(f"Number of publications in 2015-2018: {c} ({c*100/total_pubs:.1f})")
print(f"Number of publications in 2019-2022: {d} ({d*100/total_pubs:.1f})")

a = df[df.year.isin([2007, 2008, 2009, 2010])].researcher_id.nunique()
b = df[df.year.isin([2011, 2012, 2013, 2014])].researcher_id.nunique()
c = df[df.year.isin([2015, 2016, 2017, 2018])].researcher_id.nunique()
d = df[df.year.isin([2019, 2020, 2021, 2022])].researcher_id.nunique()

total_authors = df.researcher_id.nunique()

# sum of publications in bins of 4 years
print(f"Number of authors in 2007-2010: {a} ({a*100/total_authors:.1f})")
print(f"Number of authors in 2011-2014: {b} ({b*100/total_authors:.1f})")
print(f"Number of authors in 2015-2018: {c} ({c*100/total_authors:.1f})")
print(f"Number of authors in 2019-2022: {d} ({d*100/total_authors:.1f})")

a = df[df.journal == 'The Lancet'].researcher_id.nunique()
b = df[df.journal == 'JAMA'].researcher_id.nunique()
c = df[df.journal == 'The BMJ'].researcher_id.nunique()
d = df[df.journal == 'Nature Medicine'].researcher_id.nunique()
e = df[df.journal == 'New England Journal of Medicine'].researcher_id.nunique()

total_authors = df.researcher_id.nunique()

# sum of publications in bins of 4 years
print(f"Number of authors in The Lancet: {a} ({a*100/total_authors:.1f})")
print(f"Number of authors in JAMA: {b} ({b*100/total_authors:.1f})")
print(f"Number of authors in 2015-2018: {c} ({c*100/total_authors:.1f})")
print(f"Number of authors in 2019-2022: {d} ({d*100/total_authors:.1f})")
print(f"Number of authors in 2019-2022: {e} ({e*100/total_authors:.1f})")

a = df[df.journal == 'The Lancet'].pub_id.nunique()
b = df[df.journal == 'JAMA'].pub_id.nunique()
c = df[df.journal == 'The BMJ'].pub_id.nunique()
d = df[df.journal == 'Nature Medicine'].pub_id.nunique()
e = df[df.journal == 'New England Journal of Medicine'].pub_id.nunique()

total_pubs = df.pub_id.nunique()

# sum of publications in bins of 4 years
print(f"Number of publications in The Lancet: {a} ({a*100/total_pubs:.2f})")
print(f"Number of publications in JAMA: {b} ({b*100/total_pubs:.2f})")
print(f"Number of publications in 2015-2018: {c} ({c*100/total_pubs:.2f})")
print(f"Number of publications in 2019-2022: {d} ({d*100/total_pubs:.2f})")
print(f"Number of publications in 2019-2022: {e} ({e*100/total_pubs:.2f})")

# print the number of publications per journal
jf = df.groupby('journal').agg({'pub_id': 'count'}).sort_values('pub_id', ascending=False)
print(jf)

# print the % of publications per journal
print(jf/jf.sum()*100)  

# number of authors per journal
jaf = df.groupby('journal').researcher_id.nunique('researcher_id').reset_index()
jaf

jaf.researcher_id/df.researcher_id.nunique()*100


import numpy as np
def percentile(n):
    def percentile_(x):
        return np.percentile(x, n)
    percentile_.__name__ = 'percentile_%s' % n
    return percentile_

# get the median and IQR of the number of authors per publication per journal
apf = df.groupby(['pub_id', 'journal']) \
         .agg({'researcher_id': 'count'}) \
            .reset_index() \
            .groupby('journal') \
            .agg({'researcher_id': ['median', percentile(25), percentile(75)]}) \
            .reset_index()

apf

# create new column in df with the year bins
df['year_bin'] = df['year'].apply(lambda x: '2007-2010' if x in [2007, 2008, 2009, 2010] else \
                                             '2011-2014' if x in [2011, 2012, 2013, 2014] else \
                                                '2015-2018' if x in [2015, 2016, 2017, 2018] else \
                                                    '2019-2022')

# get the median and IQR of the number of authors per publication per bin of 4 years
apyf = df.groupby(['pub_id', 'year_bin']) \
            .agg({'researcher_id': 'count'}) \
            .reset_index() \
            .groupby('year_bin') \
            .agg({'researcher_id': ['median', percentile(25), percentile(75)]}) \
            .reset_index()

apyf


# print number of LMIC authors
print(df.LMIC.value_counts())
print(df.LMIC.value_counts(normalize=True))

# print number of publications with at least one LMIC author
print(df.groupby('pub_id').agg({'LMIC': 'max'}).LMIC.value_counts())
print(df.groupby('pub_id').agg({'LMIC': 'max'}).LMIC.value_counts(normalize=True))

# print number of female authors with value counts
print(df.gender.value_counts())
print(df.gender.value_counts(normalize=True))

# print publications with at least one female author
print(df.groupby('pub_id').agg({'gender': 'max'}).gender.value_counts())
print(df.groupby('pub_id').agg({'gender': 'max'}).gender.value_counts(normalize=True))

# get number of authors per year bin
ybaf = df.groupby('year_bin').agg({'researcher_id': 'count'}).reset_index()
ybaf

# same now normalized by the number of authors in total
ybaf['researcher_id']/ybaf.researcher_id.sum()*100


# number of authors per journal
jaf = df.groupby('journal').agg({'researcher_id': 'count'}).reset_index()
jaf

df.to_csv('../data_light/data_clean.csv', index=False)

df.head()

# create tableone
