==> oligarchies/3_cliques.ipynb <==
import pandas as pd
import networkx as nx
from tqdm import tqdm
import warnings
warnings.filterwarnings('ignore')

# Read the graph
G = nx.read_graphml("graphs/all.graphml")

# Find the maximum cliques
cliques = list(nx.find_cliques(G))

# print the number of cliques
print(f"Number of cliques: {len(cliques)}")
# print the size of the largest clique
print(f"Size of the largest clique: {len(max(cliques, key=len))}")

# load data_clean
df = pd.read_csv("../data_light/data_clean.csv")
df['researchers_per_pub'] = df.groupby('pub_id')['researcher_id'].transform('count')

# for all cliques, compute the number of publications
publications_per_clique = []
for clique in tqdm(cliques):
    clique_publications = df[df['researcher_id'].isin(clique)]['pub_id'].unique().tolist()
    # number of publications per clique
    publications_per_clique.append(len(clique_publications))    


# create a DataFrame to store the results
clique_df = pd.DataFrame({
    'clique': cliques,
    'publications': publications_per_clique,
    'authors': [len(clique) for clique in cliques]
}).sort_values(by='publications', ascending=False)

# save the DataFrame to a CSV file
clique_df.to_csv("cliques/cliques.csv", index=False)


# Top 5 cliques per number of publications
# find 5 cliques that do not intersect, start with cliques with most publications, stop when 5 cliques are found
clique_df_sorted = clique_df.sort_values(by=['publications'], ascending=False)

selected_cliques = []
most_pubs_indexes = []

for index, row in tqdm(clique_df_sorted.iterrows()):
    clique = set(row['clique'])

    if all(len(clique.intersection(selected_clique)) == 0 for selected_clique in selected_cliques):
        selected_cliques.append(clique)
        most_pubs_indexes.append(index)

    if len(selected_cliques) == 5:
        break

# iterate over the selected cliques, find the researchers, save to dataframe the 
# number of authors, number of publications, percentages of LMIC, percentages of female
clique_df_selected = pd.DataFrame()

for clique in tqdm(selected_cliques):
    clique_df_selected = clique_df_selected.append({
        'authors': len(clique),
        'publications': len(df[df['researcher_id'].isin(clique)]['pub_id'].unique()),
        'HIC': df[df['researcher_id'].isin(clique)]['LMIC'].value_counts(normalize=True)[0],
        'male': df[df['researcher_id'].isin(clique)]['gender'].value_counts(normalize=True)[0]
    }, ignore_index=True)


clique_df_selected

# Top 5 Cliques by number of authors
# find 5 cliques that do not intersect, start with cliques with most authors, stop when 5 cliques are found
clique_df_sorted = clique_df.sort_values(by=['authors'], ascending=False)

selected_cliques = []
most_pubs_indexes = []

for index, row in tqdm(clique_df_sorted.iterrows()):
    clique = set(row['clique'])

    if all(len(clique.intersection(selected_clique)) == 0 for selected_clique in selected_cliques):
        selected_cliques.append(clique)
        most_pubs_indexes.append(index)

    if len(selected_cliques) == 5:
        break

# iterate over the selected cliques, find the researchers, save to dataframe the 
# number of authors, number of publications, percentages of LMIC, percentages of female
clique_df_selected = pd.DataFrame()

for clique in tqdm(selected_cliques):
    clique_df_selected = clique_df_selected.append({
        'authors': len(clique),
        'publications': len(df[df['researcher_id'].isin(clique)]['pub_id'].unique()),
        'HIC': df[df['researcher_id'].isin(clique)]['LMIC'].value_counts(normalize=True)[0],
        'male': df[df['researcher_id'].isin(clique)]['gender'].value_counts(normalize=True)[0]
    }, ignore_index=True)


clique_df_selected
